doc comments and organize
