+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/Config.java
@@ -4,7 +4,7 @@ package dev.openfeature.contrib.providers.flagd;
  * Helper class to hold configuration default values.
  */
 public final class Config {
-    static final Resolver DEFAULT_RESOLVER = Resolver.flagd;
+    static final ResolverType DEFAULT_RESOLVER_TYPE = ResolverType.flagd;
     static final String DEFAULT_PORT = "8013";
     static final String DEFAULT_TLS = "false";
     static final String DEFAULT_HOST = "localhost";
@@ -51,8 +51,15 @@ public final class Config {
         }
     }
 
-    public enum Resolver {
+    public enum ResolverType {
+        /**
+         * This is the default resolver type, which connects to flagd instance with flag evaluation gRPC contract.
+         */
         flagd,
+        /**
+         * This is the in-process resolving type, where flags are fetched with flag sync gRPC contract and stored
+         * locally for in-process evaluation.
+         */
         inProcess
     }
 }
