+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/model/StringSerializer.java
@@ -7,6 +7,9 @@ import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 
 import java.io.IOException;
 
+/**
+ * Custom serializer to preserve Json node as a {@link String}
+ * */
 class StringSerializer extends StdDeserializer<String> {
 
     public StringSerializer() {
