+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/storage/FlagStore.java
@@ -1,10 +1,8 @@
 package dev.openfeature.contrib.providers.flagd.resolver.process.storage;
 
-import dev.openfeature.contrib.providers.flagd.FlagdOptions;
 import dev.openfeature.contrib.providers.flagd.resolver.process.model.FeatureFlag;
 import dev.openfeature.contrib.providers.flagd.resolver.process.model.FlagParser;
 import dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.Connector;
-import dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.GrpcStreamConnector;
 import dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.StreamPayload;
 import lombok.extern.java.Log;
 
@@ -17,14 +15,14 @@ import java.util.logging.Level;
 
 @Log
 public class FlagStore implements Storage {
+    private final Object sync = new Object();
     private final Map<String, FeatureFlag> flags = new ConcurrentHashMap<>();
-
     private final AtomicBoolean shutdown = new AtomicBoolean(false);
-    private final FlagParser parser = new FlagParser();
+
     private final Connector connector;
 
-    public FlagStore(final FlagdOptions options) {
-        this.connector = new GrpcStreamConnector(options);
+    public FlagStore(final Connector connector) {
+        this.connector = connector;
     }
 
     public void init() {
@@ -46,22 +44,26 @@ public class FlagStore implements Storage {
     }
 
     public FeatureFlag getFLag(final String key) {
-        return flags.get(key);
+        synchronized (sync) {
+            return flags.get(key);
+        }
     }
 
     private void streamerListener(final Connector connector) throws InterruptedException {
         final BlockingQueue<StreamPayload> streamPayloads = connector.getStream();
-        while (!shutdown.get()) {
-            StreamPayload take = streamPayloads.take();
 
+        while (!shutdown.get()) {
+            final StreamPayload take = streamPayloads.take();
             switch (take.getType()) {
                 case Data:
                     try {
-                        Map<String, FeatureFlag> flagMap = parser.parseString(take.getData());
-                        flags.clear();
-                        flags.putAll(flagMap);
+                        Map<String, FeatureFlag> flagMap = FlagParser.parseString(take.getData());
+                        synchronized (sync) {
+                            flags.clear();
+                            flags.putAll(flagMap);
+                        }
                     } catch (IOException e) {
-                        log.log(Level.WARNING, "Invalid flag sync payload", e);
+                        log.log(Level.WARNING, "Invalid flag sync payload from connector", e);
                     }
                 case Error:
                     // todo - event handling
