+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/storage/connector/Connector.java
@@ -2,6 +2,10 @@ package dev.openfeature.contrib.providers.flagd.resolver.process.storage.connect
 
 import java.util.concurrent.BlockingQueue;
 
+/**
+ * Contract of the in-process storage connector. Connectors are responsible to stream flag configurations in
+ * {@link StreamPayload} format.
+ */
 public interface Connector {
     void init();
