+++ providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/storage/connector/grpc/GrpcResponseModel.java
@@ -1,4 +1,4 @@
-package dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector;
+package dev.openfeature.contrib.providers.flagd.resolver.process.storage.connector.grpc;
 
 import lombok.Getter;
 import sync.v1.SyncService;
